Toronto Fitness Club Backend/PB-group_8429/requirements.txt
django==4.1.3
Pillow==9.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
django-extensions==3.2.1
django-filter==22.1
geocoder
haversine
python-dateutil